Back/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pydantic[email]==1.10.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
celery==5.2.7
redis==4.5.4
python-dotenv==1.0.0
python-multipart==0.0.9
#Potom
transliterate==1.10.2
alembic==1.14.1
MarkupSafe==3.0.2
Mako==1.3.9
httptools==0.6.4
pyOpenSSL==25.0.0
pyyaml==6.0.2 
watchfiles==1.0.4 
websockets==15.0
#eshyo
# Jinja2==3.1.5 
# aiosmtplib==3.0.2 
# blinker==1.9.0 
# fastapi-mail==1.4.2 
# pydantic==2.10.6 
# pydantic-core==2.27.2 
# pydantic-settings==2.8.1